remake devextreme-vue generator for TS type inference
+ refactor files struct
